+++ CHANGELOG.md
@@ -8,6 +8,13 @@
 ## [0.1.0] - 01/11/2025  
 
 ### Added
+ - Logic to Register an external collaborator.
+ - Logic to delete an external collaborator.
+ - Sequence diagram for User Story of deliting an external collaborator.
+ - Sequence diagram for User Story of viewing external collaborator and all of them.
+ - Logic to have the routes to view external collaborators (all and by ID).
+ - Use case to fetch all supply batches with the MVC structure (supplyBatch)
+ - Use case to fetch one supply batch with the MVC structure (supplyBatch)
  - Modified supplies model to correctly fetch data from the database (search, order, filter)
  - Updated supplies controller to handle errors when fetching supplies by ID.
  - Method to handle post request in index.js.